pravka

+++ src/BmiService.java
@@ -1,9 +1,9 @@
 public class BmiService {
-    public double calculate(double heightInMeters, double weightInKilograms) {
+    public int calculate(double heightInMeters, double weightInKilograms) {
 
         double index;
         index = weightInKilograms / (heightInMeters * heightInMeters);
 
-        return index;
+        return (int)index;
     }
 }
